refactor: consolidate TrafficLightTree and TrafficLightIndex
